+++ crates/matrix-sdk-crypto/src/types/events/to_device/mod.rs
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+//! Types for to-device events.
+
 use std::{collections::BTreeMap, fmt::Debug};
 
 use ruma::{
@@ -38,14 +40,15 @@ use zeroize::Zeroize;
 use super::{
     dummy::DummyEvent,
     forwarded_room_key::{ForwardedRoomKeyContent, ForwardedRoomKeyEvent},
-    room::encrypted::EncryptedToDeviceEvent,
     room_key::RoomKeyEvent,
     room_key_request::RoomKeyRequestEvent,
     room_key_withheld::RoomKeyWithheldEvent,
     secret_send::SecretSendEvent,
     EventType,
 };
-use crate::types::events::from_str;
+use crate::types::events::{from_str, to_device::encrypted::EncryptedToDeviceEvent};
+
+pub mod encrypted;
 
 /// An enum over the various to-device events we support.
 #[derive(Debug)]
@@ -540,7 +543,7 @@ mod tests {
             dummy_event => Dummy,
 
             // `m.room.encrypted`
-            crate::types::events::room::encrypted::tests::to_device_json => RoomEncrypted,
+            crate::types::events::to_device::encrypted::tests::json => RoomEncrypted,
         );
 
         Ok(())
@@ -587,7 +590,7 @@ mod tests {
             dummy_event => Dummy,
 
             // `m.room.encrypted`
-            crate::types::events::room::encrypted::tests::to_device_json => RoomEncrypted,
+            crate::types::events::to_device::encrypted::tests::json => RoomEncrypted,
         );
 
         Ok(())